move selected product data into PDP store
