+++ src/lib/stores/index.js
@@ -1,5 +1,6 @@
 import { configure } from "mobx";
 import AuthStore from "./AuthStore";
+import PdpStore from "./PdpStore";
 import RoutingStore from "./RoutingStore";
 import UIStore from "./UIStore";
 
@@ -7,6 +8,7 @@ configure({ enforceActions: true });
 
 export default {
   authStore: new AuthStore(),
+  pdpStore: new PdpStore(),
   routingStore: new RoutingStore(),
   uiStore: new UIStore()
 };
